# backend/internal/api/model/types.go
package model

import "github.com/blockorol/focus_flow/backend/internal/api/generated"

type AuthResponse = generated.AuthResponse
type AuthSession = generated.AuthSession
type CreateFlowRequest = generated.CreateFlowRequest
type CreateFocusRequest = generated.CreateFocusRequest
type CreateGoalRequest = generated.CreateGoalRequest
type ErrorResponse = generated.ErrorResponse
type Focus = generated.Focus
type FocusPage = generated.FocusPage
type FocusResponse = generated.FocusResponse
type Goal = generated.Goal
type GoalPage = generated.GoalPage
type GoalProgress = generated.GoalProgress
type GoalResponse = generated.GoalResponse
type LinkGoalFocusRequest = generated.LinkGoalFocusRequest
type LoginRequest = generated.LoginRequest
type UpdateFocusRequest = generated.UpdateFocusRequest
type UpdateGoalRequest = generated.UpdateGoalRequest
type User = generated.User